agent: Fix unnecessary_cast warnings
Replace 'as u8' casts with type suffix literals (_u8) for binary
literals to fix clippy warnings in Rust 1.93.

+++ src/agent/rustjail/src/cgroups/systemd/subsystem/pids.rs
@@ -53,7 +53,7 @@ mod tests {
     #[test]
     fn test_subsystem_workflow() {
         let mut pids_resources = oci::LinuxPids::default();
-        pids_resources.set_limit(0 as i64);
+        pids_resources.set_limit(0_i64);
 
         let mut properties: Properties = vec![];
